Move server functionality to new file
* Minimize code in server.rs to only start a new server, while moving
server functionality to lib.rs

* Remove some duplicate code
