+++ RelayProofsOfConcept/EddsaTokioServer/.gitignore
@@ -1,0 +1,15 @@
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+common/target/
+
+# Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
+# More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
+Cargo.lock
+
+# These are backup files generated by rustfmt
+**/*.rs.bk
+
+.idea
+.DS_Store
+.*.swp
