+++ RelayProofsOfConcept/EddsaTokioServer/Cargo.toml
@@ -9,6 +9,7 @@ serde_derive = "1.0"
 serde_json = "1.0"
 futures = "0.1"
 tokio-core = "0.1"
+tokio-io = "0.1"
 byteorder = "1.0"
 dict = "0.1.5"
 chrono = "0.4"
@@ -23,4 +24,6 @@ multi-party-ed25519 = { git = "https://github.com/KZen-networks/multi-party-edds
 git = "https://github.com/KZen-networks/curv"
 features=["ec_ed25519"]
 
-
+[lib]
+name = "relay_server"
+path = "src/lib.rs"
